requirements.txt
azure_storage==0.37.0
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.47.1
streamlit==1.28.1
